requirements.txt
#pip install -r requirements.txt

#text
transformers
captum
sentencepiece
accelerate

# Core deep learning
tensorflow==2.16.1

# (Optional) PyTorch stack – keep only if you use PyTorch elsewhere
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# Audio processing
librosa==0.10.2.post1
soundfile==0.12.1
scipy==1.13.1
numpy==1.26.4

# Machine learning & metrics
scikit-learn==1.5.1

# Visualization
matplotlib==3.9.0
seaborn==0.13.2

# Utilities
tqdm==4.66.4
pandas==2.2.2

# Optional (interactive dev)
jupyter==1.0.0

#video
timm
opencv-python
mtcnn
pytorch-grad-cam
Pillow

PyPDF2 
python-docx 
textract